improve support of file system modules (#170)

+++ cloudify_tf/utils.py
@@ -305,21 +305,23 @@ def update_terraform_source_material(new_source, target=False):
         new_source_location, dir=node_instance_dir,
         username=new_source_username,
         password=new_source_password)
-    ctx.logger.debug('Source Temp Path {}'.format(source_tmp_path))
     # check if we actually downloaded something or not
     if source_tmp_path == new_source_location:
         source_tmp_path = _create_source_path(source_tmp_path)
+    ctx.logger.debug('Source Temp Path {}'.format(source_tmp_path))
     # By getting here we will have extracted source
     # Zip the file to store in runtime
-    terraform_source_zip = _zip_archive(source_tmp_path)
-    bytes_source = _file_to_base64(terraform_source_zip)
-    os.remove(terraform_source_zip)
+    if not v1_gteq_v2(get_cloudify_version(), "6.0.0"):
 
-    if os.path.abspath(os.path.dirname(source_tmp_path)) != \
-            os.path.abspath(get_node_instance_dir(target)):
-        remove_dir(source_tmp_path)
+        terraform_source_zip = _zip_archive(source_tmp_path)
+        bytes_source = _file_to_base64(terraform_source_zip)
+        os.remove(terraform_source_zip)
 
-    return bytes_source
+        if os.path.abspath(os.path.dirname(source_tmp_path)) != \
+                os.path.abspath(get_node_instance_dir(target)):
+            remove_dir(source_tmp_path)
+
+        return bytes_source
 
 
 def get_terraform_source_material(target=False):
@@ -546,7 +548,14 @@ def update_terraform_source(new_source=None, new_source_path=None):
         # If the plan operation passes NOone, then this would error.
         material = get_terraform_source_material()
     module_root = get_storage_path()
-    extract_binary_tf_data(module_root, material, new_source_path)
+    if material:
+        extract_binary_tf_data(module_root, material, new_source_path)
+    else:
+        if isinstance(new_source, str) and os.path.isdir(new_source):
+            return new_source
+        elif isinstance(new_source, dict) and os.path.isdir(
+                new_source.get('location')):
+            return new_source.get('location')
     ctx.logger.debug('The storage root tree:\n{}'.format(tree(module_root)))
     return get_node_instance_dir(source_path=new_source_path)
 
@@ -558,9 +567,20 @@ def _yield_terraform_source(material, source_path=None):
     """
     module_root = get_storage_path()
     source_path = source_path or get_source_path()
-    extract_binary_tf_data(module_root, material, source_path)
-    ctx.logger.debug('The storage root tree:\n{}'.format(tree(module_root)))
-    path_to_init_dir = get_node_instance_dir(source_path=source_path)
+    path_to_init_dir = None
+    if not material:
+        source = get_resource_config().get('source')
+        if isinstance(source, str) and os.path.isdir(source):
+            path_to_init_dir = source
+        elif isinstance(source, dict):
+            location = source.get('location')
+            if os.path.isdir(location):
+                path_to_init_dir = source
+    if not path_to_init_dir:
+        extract_binary_tf_data(module_root, material, source_path)
+        ctx.logger.debug(
+            'The storage root tree:\n{}'.format(tree(module_root)))
+        path_to_init_dir = get_node_instance_dir(source_path=source_path)
     try:
         yield path_to_init_dir
     except Exception as e:
@@ -623,6 +643,9 @@ def get_terraform_state_file(target_dir=None):
     terraform.refresh_resources operations and it's possible we can
     get rid of it.
     """
+
+    if isinstance(target_dir, dict):
+        target_dir = target_dir.get('location')
 
     for p in os.listdir(target_dir):
         if p.endswith('tfstate'):
@@ -755,6 +778,8 @@ def handle_previous_source_format(source):
     ctx.logger.info('Source: {}'.format(source))
     if isinstance(source, dict):
         return source
+    elif isinstance(source, str) and source.startswith('/'):
+        return {'location': source}
     elif isinstance(source, str) and is_url(source):
         return {'location': source}
     try:
